quick fix lol

+++ src/kernel/string.c
@@ -5,7 +5,7 @@ void* memcpy(void* dst, const void* src, size_t num)
     uint8_t* u8Dst = (uint8_t *)dst;
     const uint8_t* u8Src = (const uint8_t *)src;
 
-    for (uint16_t i = 0; i < num; i++)
+    for (size_t i = 0; i < num; i++)
         u8Dst[i] = u8Src[i];
 
     return dst;
@@ -15,7 +15,7 @@ void * memset(void * ptr, int value, size_t num)
 {
     uint8_t* u8Ptr = (uint8_t *)ptr;
 
-    for (uint16_t i = 0; i < num; i++)
+    for (size_t i = 0; i < num; i++)
         u8Ptr[i] = (uint8_t)value;
 
     return ptr;
@@ -26,7 +26,7 @@ int memcmp(const void* ptr1, const void* ptr2, size_t num)
     const uint8_t* u8Ptr1 = (const uint8_t *)ptr1;
     const uint8_t* u8Ptr2 = (const uint8_t *)ptr2;
 
-    for (uint16_t i = 0; i < num; i++)
+    for (size_t i = 0; i < num; i++)
         if (u8Ptr1[i] != u8Ptr2[i])
             return 1;
